requirements.txt

altair==5.5.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
narwhals==2.12.0
numpy==2.3.5
packaging==25.0
pandas==2.3.3
patsy==1.0.2
pillow==12.0.0
plotly==6.5.0
plotly-express==0.4.1
protobuf==6.33.1
psycopg2-binary==2.9.11
pyarrow==21.0.0
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.29.0
scikit-learn==1.7.2
scipy==1.16.3
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.44
statsmodels==0.14.5
streamlit==1.51.0
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
